requirements.txt
gradio
ultralytics
google-generativeai
python-dotenv
opencv-python